// packages/core/src/generated/CreateSessionResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type CreateSessionResponse = { sessionId: string, sessionSecret: string, shareUrl: string, publisherToken: string, livekitUrl: string, 
/**
 * Present iff the stream is closed. Speak it to your friends.
 */
pin: string | null, };